Rename VisualNovel component
